cli/chainSpec: Move to dedicated module
